renderer_opengl: implement layer stack composition
